+++ app/jobs/cache/preheating_job.rb
@@ -5,7 +5,11 @@ class Cache::PreheatingJob < ApplicationJob
 
   def perform
     User.find_each do |user|
-      Rails.cache.write("dawarich/user_#{user.id}_years_tracked", user.years_tracked, expires_in: 1.day)
+      Rails.cache.write(
+        "dawarich/user_#{user.id}_years_tracked",
+        user.years_tracked,
+        expires_in: 1.day
+      )
     end
   end
 end